requirements.txt
streamlit
sentence-transformers
PyPDF2
pinecone-client==4.1.0
pandas
python-dotenv